+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] 
+
+### Added
+
+- Builder class for Properties
+
+### Changed
+
+- InitMocks in tests now uses the new configuration
+
+### Removed
+
+- Public visibility of _configure_ method in Properties class
+
 ## [1.3.2] 
 
 ### Changed